requirements.txt
absl-py==1.4.0
alembic==1.14.1
altgraph==0.17.4
array_record==0.5.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
auto-py-to-exe==2.45.1
avro-python3==1.10.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.36.11
botocore==1.36.11
bottle==0.13.2
bottle-websocket==0.2.9
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.9.0
contextlib2==21.6.0
contourpy==1.3.0
crcmod==1.7
cryptography==3.4.8
cycler==0.12.1
Cython==3.0.11
dill==0.3.1.1
distro==1.9.0
dm-tree==0.1.8
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
Eel==0.18.1
etils==1.9.1
fastavro==1.9.5
fasteners==0.19
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.53.1
fredapi==0.5.2
frozendict==2.4.6
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
gevent==24.11.1
gevent-websocket==0.10.1
gin-config==0.5.0
google-api-core==2.19.2
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpcio==1.66.0
h5py==3.11.0
hdfs==2.7.3
html5lib==1.1
httplib2==0.20.2
idna==3.8
immutabledict==4.2.0
importlib-metadata==4.6.4
importlib_resources==6.4.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
Js2Py==0.74
jsonpickle==3.2.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaggle==1.6.17
keras==3.5.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lvis==0.5.3
lxml==5.3.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==8.10.0
multitasking==0.0.11
namex==0.0.8
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.0
objsize==0.7.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
optuna==4.2.1
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pandas_ta==0.3.14b0
pdfkit==1.0.0
peewee==3.17.8
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.3.6
portalocker==2.10.1
promise==2.3
proto-plus==1.24.0
protobuf==4.25.6
psutil==6.0.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydot==1.4.2
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pyjsparser==2.7.1
PyJWT==2.3.0
pymongo==4.8.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
s3transfer==0.11.2
sacrebleu==2.2.0
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.14.1
SecretStorage==3.3.1
sentencepiece==0.2.0
seqeval==1.2.2
setuptools==75.8.0
simple_parsing==0.1.5
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.34
ta==0.11.0
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-hub==0.16.1
tensorflow-io==0.37.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.15.0
tensorflow-model-optimization==0.8.0
termcolor==2.4.0
text-unidecode==1.3
tf-slim==1.1.0
tf_keras==2.18.0
threadpoolctl==3.5.0
toml==0.10.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.4
wheel==0.45.1
wrapt==1.16.0
WTForms==3.1.2
yfinance==0.2.44
zipp==1.0.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
